Code Refactoring: Updating to a separate endpoint for document generation tasks Moving test cases, unit and functional
