+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/functionaltest/servicejourney/MakeServiceDecisionTest.java
@@ -6,28 +6,41 @@ import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.ResultMatcher;
 import uk.gov.hmcts.reform.bsp.common.model.document.CtscContactDetails;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdStates;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CcdCallbackRequest;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CcdCallbackResponse;
+import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.DocumentLink;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ApplicationServiceTypes;
+import uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ServiceApplicationRefusalOrder;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.document.ServiceDecisionOrder;
 import uk.gov.hmcts.reform.divorce.orchestration.functionaltest.IdamTestSupport;
 import uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.CtscContactDetailsDataProviderService;
 import uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.DatesDataExtractor;
 import uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.DeemedServiceOrderGenerationTask;
+import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.DeemedServiceRefusalOrderTask;
+import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.DispensedServiceRefusalOrderTask;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.OrderToDispenseGenerationTask;
+import uk.gov.hmcts.reform.divorce.utils.DateUtils;
 
+import java.time.LocalDate;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import static com.jayway.jsonpath.matchers.JsonPathMatchers.hasJsonPath;
+import static com.jayway.jsonpath.matchers.JsonPathMatchers.hasNoJsonPath;
+import static com.jayway.jsonpath.matchers.JsonPathMatchers.isJson;
 import static java.time.LocalDate.now;
 import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
 import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.Matchers.allOf;
+import static org.hamcrest.Matchers.hasSize;
+import static org.hamcrest.Matchers.is;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
@@ -43,17 +56,24 @@ import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.R
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.SERVICE_APPLICATION_DECISION_DATE;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.SERVICE_APPLICATION_GRANTED;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.SERVICE_APPLICATION_TYPE;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdStates.SERVICE_APPLICATION_NOT_APPROVED;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.CASE_ID_JSON_KEY;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.D8DOCUMENTS_GENERATED;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.DOCUMENT_CASE_DETAILS_JSON_KEY;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.YES_VALUE;
+import static uk.gov.hmcts.reform.divorce.orchestration.functionaltest.servicejourney.ServiceDecisionMadeTest.buildRefusalRequest;
+import static uk.gov.hmcts.reform.divorce.orchestration.functionaltest.servicejourney.ServiceDecisionMadeTest.buildServiceRefusalOrderCaseData;
+import static uk.gov.hmcts.reform.divorce.orchestration.functionaltest.servicejourney.ServiceDecisionMadeTest.generateDocumentLink;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.CaseDataExtractor.CaseDataKeys.CASE_REFERENCE;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.CaseDataExtractor.getCaseReference;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.CaseDataKeys.PETITIONER_FIRST_NAME;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.CaseDataKeys.PETITIONER_LAST_NAME;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.CaseDataKeys.RESPONDENT_FIRST_NAME;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.CaseDataKeys.RESPONDENT_LAST_NAME;
+import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.getPetitionerFullName;
 import static uk.gov.hmcts.reform.divorce.orchestration.testutil.CaseDataTestHelper.createCollectionMemberDocumentAsMap;
 import static uk.gov.hmcts.reform.divorce.orchestration.testutil.ObjectMapperTestUtil.convertObjectToJsonString;
+import static uk.gov.hmcts.reform.divorce.orchestration.util.ServiceApplicationRefusalHelper.getServiceApplicationRefusalReason;
 import static uk.gov.hmcts.reform.divorce.utils.DateUtils.formatDateFromLocalDate;
 
 public class MakeServiceDecisionTest extends IdamTestSupport {
@@ -124,6 +144,51 @@ public class MakeServiceDecisionTest extends IdamTestSupport {
             .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
             .andExpect(content().json(convertObjectToJsonString(expectedResponse)));
+    }
+
+    @Test
+    public void shouldPopulateDocumentsWithDeemedRefusalOrderAndRemoveDraftWhenSubmitted() throws Exception {
+        String templateId = DeemedServiceRefusalOrderTask.FileMetadata.TEMPLATE_ID;
+        String documentType = DeemedServiceRefusalOrderTask.FileMetadata.DOCUMENT_TYPE;
+
+        CcdCallbackRequest ccdCallbackRequest = buildServiceRefusalOrderFixture(
+            templateId, ApplicationServiceTypes.DEEMED, documentType
+        );
+
+        webClient.perform(post(API_URL)
+            .contentType(MediaType.APPLICATION_JSON)
+            .header(AUTHORIZATION, AUTH_TOKEN)
+            .content(convertObjectToJsonString(ccdCallbackRequest)))
+            .andExpect(status().isOk())
+            .andExpect(commonExpectationsForServiceRefusalOrder(documentType));
+    }
+
+    @Test
+    public void shouldPopulateDocumentsWithDispensedRefusalOrderAndRemoveDraftWhenSubmitted() throws Exception {
+        String templateId = DispensedServiceRefusalOrderTask.FileMetadata.TEMPLATE_ID;
+        String documentType = DispensedServiceRefusalOrderTask.FileMetadata.DOCUMENT_TYPE;
+
+        CcdCallbackRequest ccdCallbackRequest = buildServiceRefusalOrderFixture(
+            templateId, ApplicationServiceTypes.DISPENSED, documentType
+        );
+
+        webClient.perform(post(API_URL)
+            .contentType(MediaType.APPLICATION_JSON)
+            .header(AUTHORIZATION, AUTH_TOKEN)
+            .content(convertObjectToJsonString(ccdCallbackRequest)))
+            .andExpect(status().isOk())
+            .andExpect(commonExpectationsForServiceRefusalOrder(documentType));
+    }
+
+    private ResultMatcher commonExpectationsForServiceRefusalOrder(String documentType) {
+        return content().string(allOf(
+            isJson(),
+            hasNoJsonPath("$.data.ServiceRefusalDraft"),
+            hasJsonPath("$.data.D8DocumentsGenerated", hasSize(1)),
+            hasJsonPath("$.data.D8DocumentsGenerated[0].value.DocumentType",
+                is(documentType)),
+            hasNoJsonPath("$.errors")
+        ));
     }
 
     private Map<String, Object> buildInputCaseData(String applicationType) {
@@ -210,6 +275,40 @@ public class MakeServiceDecisionTest extends IdamTestSupport {
 
     private CcdCallbackRequest buildRequest() {
         return buildRequest(new HashMap<>());
+    }
+
+    private CcdCallbackRequest buildServiceRefusalOrderFixture(
+        String templateId, String serviceType, String documentType) {
+        DocumentLink refusalDraftDocument = generateDocumentLink(templateId);
+        Map<String, Object> refusalOrderData = buildServiceRefusalOrderCaseData(
+            serviceType, refusalDraftDocument
+        );
+
+        CcdCallbackRequest ccdCallbackRequest = buildRefusalRequest(refusalOrderData);
+        ccdCallbackRequest.getCaseDetails().setState(SERVICE_APPLICATION_NOT_APPROVED);
+
+        stubDocumentGeneratorServiceBaseOnContextPath(
+            templateId,
+            singletonMap(DOCUMENT_CASE_DETAILS_JSON_KEY,
+                ImmutableMap.of(
+                    "id", refusalOrderData.get(CASE_ID_JSON_KEY),
+                    "case_data", buildPopulatedServiceRefusalOrderTemplateModel(refusalOrderData)
+                )),
+            documentType
+        );
+        return ccdCallbackRequest;
+    }
+
+    private ServiceApplicationRefusalOrder buildPopulatedServiceRefusalOrderTemplateModel(Map<String, Object> caseData) {
+        return ServiceApplicationRefusalOrder.serviceApplicationRefusalOrderBuilder()
+            .ctscContactDetails(ctscContactDetails)
+            .petitionerFullName(getPetitionerFullName(caseData))
+            .respondentFullName(FullNamesDataExtractor.getRespondentFullName(caseData))
+            .caseReference((String) caseData.get(CASE_ID_JSON_KEY))
+            .receivedServiceApplicationDate(DatesDataExtractor.getReceivedServiceApplicationDate(caseData))
+            .serviceApplicationRefusalReason(getServiceApplicationRefusalReason(caseData))
+            .documentIssuedOn(DateUtils.formatDateWithCustomerFacingFormat(LocalDate.now()))
+            .build();
     }
 }
