+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/workflows/servicejourney/ServiceDecisionMadeWorkflowTest.java
@@ -8,9 +8,6 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CaseDetails;
 import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.WorkflowException;
-import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.DeemedServiceRefusalOrderTask;
-import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.DispensedServiceRefusalOrderTask;
-import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.ServiceRefusalDraftRemovalTask;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.DeemedApprovedEmailTask;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.DeemedNotApprovedEmailTask;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.DispensedApprovedEmailTask;
@@ -44,15 +41,6 @@ public class ServiceDecisionMadeWorkflowTest {
     private ServiceDecisionMadeWorkflow classUnderTest;
 
     @Mock
-    private DeemedServiceRefusalOrderTask deemedServiceRefusalOrderTask;
-
-    @Mock
-    private DispensedServiceRefusalOrderTask dispensedServiceRefusalOrderTask;
-
-    @Mock
-    private ServiceRefusalDraftRemovalTask serviceRefusalDraftRemovalTask;
-
-    @Mock
     private DeemedApprovedEmailTask deemedApprovedEmailTask;
 
     @Mock
@@ -71,23 +59,18 @@ public class ServiceDecisionMadeWorkflowTest {
 
         mockTasksExecution(
             caseData,
-            deemedServiceRefusalOrderTask,
-            deemedNotApprovedEmailTask,
-            serviceRefusalDraftRemovalTask
+            deemedNotApprovedEmailTask
         );
 
         Map<String, Object> returnedData = executeWorkflow(caseDetails);
 
         verifyTasksCalledInOrder(
             returnedData,
-            deemedServiceRefusalOrderTask,
-            deemedNotApprovedEmailTask,
-            serviceRefusalDraftRemovalTask
+            deemedNotApprovedEmailTask
         );
 
         runNoTasksToSendApprovedEmails();
         verifyTasksWereNeverCalled(dispensedNotApprovedEmailTask);
-        verifyTasksWereNeverCalled(dispensedServiceRefusalOrderTask);
     }
 
     @Test
@@ -97,23 +80,18 @@ public class ServiceDecisionMadeWorkflowTest {
 
         mockTasksExecution(
             caseData,
-            dispensedServiceRefusalOrderTask,
-            dispensedNotApprovedEmailTask,
-            serviceRefusalDraftRemovalTask
+            dispensedNotApprovedEmailTask
         );
 
         Map<String, Object> returnedData = executeWorkflow(caseDetails);
 
         verifyTasksCalledInOrder(
             returnedData,
-            dispensedServiceRefusalOrderTask,
-            dispensedNotApprovedEmailTask,
-            serviceRefusalDraftRemovalTask
+            dispensedNotApprovedEmailTask
         );
 
         runNoTasksToSendApprovedEmails();
         verifyTasksWereNeverCalled(deemedNotApprovedEmailTask);
-        verifyTasksWereNeverCalled(deemedServiceRefusalOrderTask);
     }
 
     @Test
@@ -129,7 +107,6 @@ public class ServiceDecisionMadeWorkflowTest {
 
         verifyTaskWasNeverCalled(dispensedApprovedEmailTask);
         runNoTasksToSendNotApprovedEmails();
-        runNoTasksToGeneratePdfs();
     }
 
     @Test
@@ -146,7 +123,6 @@ public class ServiceDecisionMadeWorkflowTest {
 
         verifyTaskWasNeverCalled(deemedApprovedEmailTask);
         runNoTasksToSendNotApprovedEmails();
-        runNoTasksToGeneratePdfs();
     }
 
     @Test
@@ -158,21 +134,6 @@ public class ServiceDecisionMadeWorkflowTest {
         executeWorkflow(caseDetails);
 
         runNoTasksAtAll();
-    }
-
-    @Test
-    public void whenServiceDecisionMadeAndServiceApplicationIsNotGrantedAndAndTypeIsOtherDoNotGeneratePdfs() throws WorkflowException {
-        Map<String, Object> caseData = buildCaseData("someOtherValue", NO_VALUE);
-        CaseDetails caseDetails = buildCaseDetails(caseData, SERVICE_APPLICATION_NOT_APPROVED);
-
-        mockTasksExecution(caseData, serviceRefusalDraftRemovalTask);
-
-        Map<String, Object> returnedCaseData = executeWorkflow(caseDetails);
-
-        verifyTaskWasCalled(returnedCaseData, serviceRefusalDraftRemovalTask);
-
-        runNoTasksToSendEmails();
-        runNoTasksToGenerateFinalPdfs();
     }
 
     private Map<String, Object> buildCaseData(String serviceApplicationType, String serviceApplicationGranted) {
@@ -188,18 +149,6 @@ public class ServiceDecisionMadeWorkflowTest {
             .caseId(TEST_CASE_ID)
             .state(caseState)
             .build();
-    }
-
-    private void runNoTasksToGenerateFinalPdfs() {
-        verifyTasksWereNeverCalled(
-            deemedServiceRefusalOrderTask,
-            dispensedServiceRefusalOrderTask
-        );
-    }
-
-    private void runNoTasksToGeneratePdfs() {
-        runNoTasksToGenerateFinalPdfs();
-        verifyTasksWereNeverCalled(serviceRefusalDraftRemovalTask);
     }
 
     private void runNoTasksToSendEmails() {
@@ -219,7 +168,6 @@ public class ServiceDecisionMadeWorkflowTest {
     }
 
     private void runNoTasksAtAll() {
-        runNoTasksToGeneratePdfs();
         runNoTasksToSendEmails();
     }
